+++ src/main/java/com/github/victools/jsonschema/module/jackson/JacksonModule.java
@@ -83,6 +83,12 @@ public class JacksonModule implements Module {
         if (this.options.contains(JacksonOption.FLATTENED_ENUMS_FROM_JSONVALUE)) {
             generalConfigPart.withCustomDefinitionProvider(new CustomEnumJsonValueDefinitionProvider());
         }
+        if (!this.options.contains(JacksonOption.SKIP_SUBTYPE_LOOKUP)) {
+            JsonSubTypesResolver subtypeResolver = new JsonSubTypesResolver();
+            generalConfigPart.withSubtypeResolver(subtypeResolver);
+            fieldConfigPart.withTargetTypeOverridesResolver(field -> subtypeResolver.resolveSubtypes(field.getType(),
+                    field.getAnnotationConsideringFieldAndGetter(JsonSubTypes.class), field.getContext()));
+        }
     }
 
     /**