+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Look-up subtypes according to `@JsonTypeInfo` and `@JsonSubTypes` annotations:
+    - Considering `@JsonTypeInfo.include` with `WRAPPER_ARRAY`, `WRAPPER_OBJECT`, `PROPERTY`, `EXTERNAL_PROPERTY`, `EXISTING_PROPERTY`
+    - Considering `@JsonTypeInfo.use` with `NAME` (from `@JsonTypeName` or `@JsonSubTypes.Type`) and `CLASS` 
+- New `JacksonOption.SKIP_SUBTYPE_LOOKUP` to allow disabling the new subtype handling (i.e. to regain previous behaviour) if required
+
 ## [4.7.0] - 2020-03-25
 ### Changed
 - Ignore/exclude properties marked with `@JsonBackReference`
@@ -31,6 +38,7 @@ BEWARE: The `forTypesInGeneral()` API was only added to the main generator libra
 - Apply alternative field names defined in `@JsonProperty` annotations.
 - Ignore fields that are deemed to be ignored according to various `jackson-annotations` (e.g. `@JsonIgnore`, `@JsonIgnoreType`, `@JsonIgnoreProperties`) or are otherwise supposed to be excluded.
 
+[Unreleased]: https://github.com/victools/jsonschema-module-jackson/compare/v4.7.0...HEAD
 [4.7.0]: https://github.com/victools/jsonschema-module-jackson/compare/v4.5.1...v4.7.0
 [4.5.1]: https://github.com/victools/jsonschema-module-jackson/compare/v4.5.0...v4.5.1
 [4.5.0]: https://github.com/victools/jsonschema-module-jackson/compare/v4.4.0...v4.5.0